+++ packages/transaction-pay-controller/src/utils/quotes.ts
@@ -276,13 +276,13 @@ function syncTransaction({
         tx.isExternalSign = hasQuotes;
       }
 
-      const intent = tx.metamaskPay?.intent;
+      const source = tx.metamaskPay?.source;
 
       tx.metamaskPay = {
         bridgeFeeFiat: totals.fees.provider.usd,
         chainId: paymentToken?.chainId,
         isPostQuote,
-        ...(intent && { intent }),
+        ...(source && { source }),
         networkFeeFiat: totals.fees.sourceNetwork.estimate.usd,
         strategy,
         targetFiat: totals.targetAmount.usd,